refactor(user-profile): extract estimator Cluster 5 bodymap helper (WP2.1f)
Move `muscle_coverage_state` (Cluster 5, the last leaf) verbatim from the
`utils.profile_estimator` facade into a new private
`utils/_profile_estimator/bodymap.py`, per Deep Refactor Plan v3 Phase 2
(docs/user_profile/PROFILE_ESTIMATOR_CLUSTERS.md §2/§3/§4). Behavior-preserving
mechanical move — no estimator-priority, math, response-shape, typing, or
cleanup changes.

The facade re-exports `muscle_coverage_state` (`# noqa: F401`), so the public
`utils.profile_estimator` surface and every consumer (routes/user_profile.py,
utils/strength_calibration.py) are unchanged. The moved function body is
byte-identical to HEAD (verified LF-normalized), and the facade remainder equals
HEAD minus the moved block plus only the new import block.

bodymap.py depends only on the constants leaf (BODYMAP_MUSCLE_KEYS,
KEY_LIFT_LABELS, MUSCLE_TO_KEY_LIFT), core_math (epley_1rm), and the coverage
leaf (_is_lift_filled) — imported DIRECTLY from utils._profile_estimator.coverage,
not via the facade, so no facade↔bodymap cycle is introduced. It takes no
module-level strength_calibration import; the profile_estimator ⇄
strength_calibration cycle stays held solely by the facade's two function-local
lazy imports. Orchestration now stays in the facade (WP2.1f is the last leaf).

Object identity of re-exported names is preserved (re-exports, not copies); the
contract test asserts epley_1rm / match_direct_lift_key identity in both import
orders and both remain green.

pyright: no baseline edit — muscle_coverage_state is None-guarded throughout and
carried zero diagnostics in the facade, so relocating it produced 0 net-new
diagnostics (gate PASS, current 175 <= baseline 186).

Verification: contract + focused estimator/calibration suites 206 passed; full
pytest 1710 passed / 0 failed; blocking flake8 (E9,F63,F7,F82,F811,E711,E712,
F401) 0; pyright baseline gate PASS 0 net-new.

Migration note: none; behavior and public imports preserved.
